codalab/requirements/common.txt
Django==1.5.9
alembic==0.6.7
anyjson==0.3.3
argparse
azure==0.7.1
closure_linter
django-allauth==0.11.1
django-analytical==0.15.0
django-appconf==0.6
django-compressor==1.3
django-config-gen==1.0.8
# django-configurations
django-configurations==0.8
django-debug-toolbar==0.9.4
django-debug-panel==0.8.1
django-extensions>=1.1.1
django-extra-views==0.6.3
django-filter==0.7
django-guardian==1.1.1
django-haystack
django-js-reverse>=0.2.0
django-mptt==0.5.5
django-nose
django-oauth-toolkit==0.6.1
django-storages==1.1.8
django-tinymce
djangorestframework==2.3.6
mock==1.0.1
oauthlib==0.6.1
psutil==2.1.1
python-openid==2.2.5
python-ptrace==0.6.5
pytz==2013b
PyYAML==3.10
requests-oauthlib==0.3.2
requests==1.2.3
# six==1.3.0
six==1.10.0
South==1.0.2
SQLAlchemy==0.9.7
watchdog==0.8.1
wsgiref==0.1.2
django-simple-captcha==0.4.5
django-bootstrap-form==3.2
newrelic==2.54.0.41
python-memcached==1.57

supervisor==3.0
uWSGI==1.9.16
pycrypto==2.6

# For tests
selenium

# Required by codalab-cli
MySQL-python
docker-py==1.6.0
# Use sckoo@stanford.edu's patched argcomplete, until pull request has been accepted
# https://github.com/kislyuk/argcomplete/pull/118
git+https://github.com/kashizui/argcomplete.git@allow-overridden-subparsersaction
